Modify the design of the Website class
Modified the design of the Website class to be an abstract
base class for the two classes:

	- RightmoveWebsite
	- ZooplaWebsite

Therefore, the initial version of this project will support
only the Rightmove and zoopla websites.

The RightmoveWebsite subclass will override the get_location_identifier
method, while the ZooplaWebsite subclass will use the base
implementation.