Modify UML Diagram
Modified the name of the Website class to include {abstract}.
This makes it clear that Website is an Abstract Base Class.

Removed <<interface>> from the header of the RightmoveWebsite
and ZooplaWebsite classes. This was an error.

Added the get_location_identifier method to the ZooplaWebsite
class. This should have been there before as it must be overwritten.