Add unit test for Website 'parse_listings'
Added a unit test for the 'parse_listings' method of the Website
class.

Created the TestDate directory, which will contain all test data
needed for the unit tests.

Created the one_complete_listing.html test data file. This is an
extract of Rightmove html, used to test the parsing of html in the
Website class.