Define __eq__ and __hash__ Accommodation methods
Defined the __eq__ method for the Accommodation class, which
checks if two Accommodation objects are equal by comparing their
attributes.

Defined the __hash__ method for the Accommodation class, which
returns a hash value for the Accommodation object based on its
attributes.

These methods are required when comparing the Accommodation objects
in the test_Filer.py test class.